Introduced new functions, refactored structure, added two ways of interacting with the database

+++ src/main/java/org/example/flashcards/repository/CardRepository.java
@@ -6,10 +6,11 @@ import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Repository
 public class CardRepository {
-    private EntityManager entityManager;
+    private final EntityManager entityManager;
 
     public CardRepository(EntityManager entityManager) {
         this.entityManager = entityManager;
@@ -26,17 +27,23 @@ public class CardRepository {
     }
 
     @Transactional
-    public void delete(Card card) {
-        entityManager.remove(card);
+    public void deleteById(Long id) {
+        Card card = entityManager.find(Card.class, id);
+        if (card != null) {
+            entityManager.remove(card);
+        }
+    }
+
+    public Optional<Card> findById(Long id) {
+        return Optional.ofNullable(entityManager.find(Card.class, id));
     }
 
     public List<Card> findAll() {
         return entityManager.createQuery("select c from Card c", Card.class).getResultList();
     }
 
-    public Card findById(Long id) {
-        return entityManager.find(Card.class, id);
+    public List<Card> findAllSorted(String language, String order) {
+        String query = String.format("select c from Card c order by %s %s", language, order);
+        return entityManager.createQuery(query, Card.class).getResultList();
     }
-
-
 }
